Reorganized parser tests in separate files. Added "describe"
